backend/requirements.txt
Django>=4.2,<5.1
djangorestframework>=3.14
drf-spectacular>=0.27
django-cors-headers>=4.4
psycopg2-binary>=2.9
pdfplumber>=0.10
python-dotenv>=1.0
pypdf>=3.12
pytest>=8.3
pytest-django>=4.8
python-decouple>=3.8
python-magic>=0.4.27
gunicorn>=21.2
